refactor: restructure `blog` content collection
This is part 1 of the restructure of content collections.

The most important of these changes is how images are now handle; no
more cursed names, and they are side by side to their respective content
collection entry.

For part 2, I want to also apply this same scheme to "projects" and to
see if there is a way to remove the need of "image" properties by
creating `<Image />` components by using the `[id]` property of the
content collection entry.

chore: prefer not accented vowels for blog ids

chore: `episodio-1` -> `rumbo-a-pycon-us-episodio-1`